Gestion documentaire pour : Les etudiants, les etudes, les formations, les prospects.

+++ src/mgate/PubliBundle/Controller/DocumentController.php
@@ -2,13 +2,12 @@
 namespace mgate\PubliBundle\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-use Acme\DemoBundle\Entity\Document;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
 use JMS\SecurityExtraBundle\Annotation\Secure;
 
-use mgate\PubliBundle\Entity\DocumentType;
+use mgate\PubliBundle\Entity\Document;
 
-class DocumentTypeController extends Controller
+class DocumentController extends Controller
 {
     /**
      * @Secure(roles="ROLE_SUIVEUR")
@@ -17,9 +16,9 @@ class DocumentTypeController extends Controller
     {
         $em = $this->getDoctrine()->getManager();
 
-        $entities = $em->getRepository('mgatePubliBundle:DocumentType')->findAll();
+        $entities = $em->getRepository('mgatePubliBundle:Document')->findAll();
       
-        return $this->render('mgatePubliBundle:DocumentType:index.html.twig', array(
+        return $this->render('mgatePubliBundle:Document:index.html.twig', array(
             'docs' => $entities,
         ));
        
@@ -28,9 +27,9 @@ class DocumentTypeController extends Controller
     /**
      * @Secure(roles="ROLE_SUIVEUR")
      */
-    public function uploadAction()
+    public function uploadAction($deleteIfExist = false)
     {
-        $document = new DocumentType();
+        $document = new Document();
         $form = $this->createFormBuilder($document)
             ->add('name')
             ->add('file')
@@ -41,25 +40,25 @@ class DocumentTypeController extends Controller
             if ($form->isValid())
             {
                 $document->setName(strtoupper($document->getName()));
-                
                 $em = $this->getDoctrine()->getManager();
-
-                $docs = $em->getRepository('mgatePubliBundle:DocumentType')->findBy(array('name' => $document->getName() )); // Ligne qui posse problème
-                if ($docs) {
-                    foreach ($docs as $doc) {
-                        $em->remove($doc);
-                    }
-                }
-                
-                
                 $em->persist($document);
                 $em->flush();
 
+                
+                if($deleteIfExist){
+                    $docs = $em->getRepository('mgatePubliBundle:Document')->findBy(array('name' => $document->getName() ));
+                    if ($docs) {
+                        foreach ($docs as $doc) {
+                            $em->remove($doc);
+                        }
+                    }
+                }
+                
                 $this->redirect($this->generateUrl('mgate_publi_documenttype_index'));
             }
         }
 
         //return array('form' => $form->createView());
-        return $this->render('mgatePubliBundle:DocumentType:upload.html.twig', array('form' => $form->createView()));
+        return $this->render('mgatePubliBundle:Document:upload.html.twig', array('form' => $form->createView()));
     }
 }
